move get_scopes to it's own module
